Tests: Move StackLayoutTests helpers into standalone helper enum
This is related to #740
